IXIUP-2 - implemented Registered Names - partial implementation of RN APIs - added total fee to block header - Ixiac Omega improvements and fixes - partial implementation of RegNet - minor improvements to network protocol - other improvements

+++ Wallet/Wallet.cs
@@ -220,7 +220,7 @@ namespace IXICore
             }
 
             rawData.AddRange(BitConverter.GetBytes((int)type));
-            rawData.AddRange(BitConverter.GetBytes(requiredSigs));
+            rawData.AddRange(BitConverter.GetBytes((short)requiredSigs));
 
             if (allowedSigners != null)
             {
